+++ augmentations.d.ts
@@ -2,6 +2,8 @@ import messages from "@cucumber/messages";
 
 import { Registry } from "./lib/registry";
 
+import { MochaGlobals } from "mocha";
+
 declare module "@cucumber/cucumber" {
   interface IWorld {
     tmpDir: string;
@@ -24,6 +26,10 @@ declare global {
   namespace globalThis {
     var __cypress_cucumber_preprocessor_registry_dont_use_this:
       | Registry
+      | undefined;
+
+    var __cypress_cucumber_preprocessor_mocha_dont_use_this:
+      | Pick<MochaGlobals, "before" | "beforeEach" | "after" | "afterEach">
       | undefined;
   }