+++ lib/subpath-entrypoints/webpack.ts
@@ -5,7 +5,9 @@ import { compile } from "../template";
 const loader: LoaderDefinition = function (data) {
   const callback = this.async();
 
-  compile(this.query as any, data, this.resourcePath).then(
+  const config: Cypress.PluginConfigOptions = this.query as any;
+
+  compile(config, data, this.resourcePath).then(
     (result) => callback(null, result),
     (error) => callback(error),
   );